@@ -16,18 +16,18 @@
         }
     },
     "require": {
-        "php": ">=8.0.0",
-        "psy/psysh": "^0.11.9"
+        "php": ">=8.1.0",
+        "psy/psysh": "^0.11.20"
     },
     "require-dev": {
-        "phpunit/phpunit": "^9.6|^10.0",
-        "mockery/mockery": "^1.5",
-        "phpstan/phpstan": "^1.4",
-        "mako/framework": "^9.0.0-dev"
+        "phpunit/phpunit": "^10.3",
+        "mockery/mockery": "^1.6",
+        "phpstan/phpstan": "^1.10",
+        "mako/framework": "^10.0.0-dev"
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "5.0-dev"
+            "dev-master": "6.0-dev"
         }
     }
 }
